moved badge to seperate widget
